% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add1_AIC.R
\name{add1_AIC}
\alias{add1_AIC}
\alias{add1_AIC_mu}
\title{Add one possible covariate on GEV parameter based on AIC}
\usage{
add1_AIC_mu(fit)
}
\arguments{
\item{fit}{An object of class \code{c("gev", "evreg")} returned from
\code{\link{gevreg}} summarising the current model fit.}
}
\value{
An object (a list) of class \code{c("gev", "evreg")} summarising
  the new model fit (which may be the same as \code{fit}) and containing the
  following additional components
    \item{Input_fit}{The input object of the class \code{c("gev", "evreg")}.}
    \item{Note}{A message that tells if a covariate has been added or not.}
    \item{Output_fit}{A list that contains formulae for the parameter,
    and the output object of the class \code{c("gev", "evreg")} if the output fit
    is different from the input fit.}
    \item{added_covariate}{A character vector shows added covariate}
    \item{AIC}{AIC values for both input model and output model if two models
    are different.}
}
\description{
Add a single term to either mu, sigma, and xi based on AIC.
}
\details{
Add details.
}
\examples{

### Fremantle sea levels
# Parameter mu
f0 <- gevreg(SeaLevel, data = evreg::fremantle[,-1])
add1_AIC_mu(f0)

# Parameter sigma
f3 <- gevreg(y = SeaLevel, data = evreg::fremantle[,-1], mu = ~Year01 + SOI)
add1_AIC_sigma(f3)

# Parameter xi
f4 <- gevreg(y = SeaLevel, data = evreg::fremantle[,-1], mu = ~Year01 + SOI, sigma = ~SOI)
add1_AIC_xi(f4)


### Annual Maximum and Minimum Temperature

P0 <- gevreg(y = TMX1, data = PORTw[, -1])
add1_AIC_mu(P0)
add1_AIC_sigma(P0)
add1_AIC_xi(P0)

}
