% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add1_AIC.R
\name{add1_AIC_mu}
\alias{add1_AIC_mu}
\title{Add one possible covariate on GEV parameter based on AIC}
\usage{
add1_AIC_mu(fit)
}
\arguments{
\item{fit}{A model of class "gevreg".}
}
\value{
A list which has the following components
    \item{Input_fit}{The input object of the class gevreg.}
    \item{Note}{A message that will be printed when input fit and output
    fit are the same.}
    \item{Output_fit}{A list that contains formulae for the parameter,
    and the output object of the class gevreg if output fit is different
    from the input fit.}
    \item{AIC}{AIC values for both input model and output model if two models
    are different.}
}
\description{
Add a single term to either mu, sigma, and xi based on AIC.
}
\details{
add1_AIC_mu(fit)
add1_AIC_sigma(fit)
add1_AIC_xi(fit)

Add details.
}
\examples{

### Fremantle sea levels

f0 <- gevreg(SeaLevel, data = evreg::fremantle)
add1_AIC_mu(f0)

}
