% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add1_AIC.R, R/add1_LRT.R
\name{add1_AIC_sigma}
\alias{add1_AIC_sigma}
\alias{add1_AIC_xi}
\alias{add1_LRT}
\alias{add1_LRT_mu}
\alias{add1_LRT_sigma}
\alias{add1_LRT_xi}
\title{Add one possible covariate on GEV parameter based on Likelihood-ratio-test}
\usage{
add1_AIC_sigma(fit)

add1_AIC_xi(fit)

add1_LRT_mu(fit, alpha = 0.05)

add1_LRT_sigma(fit, alpha = 0.05)

add1_LRT_xi(fit, alpha = 0.05)
}
\arguments{
\item{fit}{An object of class \code{c("gev", "evreg")} returned from
\code{\link{gevreg}} summarising the current model fit.}

\item{alpha}{Significance level. Default value is 0.05.}
}
\value{
An object (a list) of class \code{c("gev", "evreg")} summarising
  the new model fit (which may be the same as \code{fit}) and containing the
  following additional components
    \item{Input_fit}{The input object of the class \code{c("gev", "evreg")}.}
    \item{Note}{A message that tells if a covariate has been added or not.}
    \item{Output_fit}{A list that contains formulae for the parameter,
    and the output object of the class \code{c("gev", "evreg")} if the output fit
    is different from the input fit.}
    \item{added_covariate}{A character vector shows added covariate}
    \item{pvalue}{A data frame that contains p value with five decimal
    places of the Likelihood-ratio-test.}
}
\description{
Add a single term to either mu, sigma, and xi based on likelihood-ratio-test.
}
\examples{
### Fremantle sea levels

f0 <- gevreg(SeaLevel, data = evreg::fremantle[, -1])
add1_LRT_mu(f0)


### Annual Maximum and Minimum Temperature

P0 <- gevreg(y = TMX1, data = PORTw[, -1])
add1_LRT_mu(P0)
add1_LRT_sigma(P0)
add1_LRT_xi(P0)

}
