% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add1_pvalue.R
\name{add1_p}
\alias{add1_p}
\alias{add1_p_mu}
\alias{add1_p_sigma}
\alias{add1_p_xi}
\title{Add one possible covariate on GEV parameter using individual p value from Wald test}
\usage{
add1_p_mu(fit, alpha = 0.05)

add1_p_sigma(fit, alpha = 0.05)

add1_p_xi(fit, alpha = 0.05)
}
\arguments{
\item{fit}{An object of class \code{c("gev", "evreg")} returned from
\code{\link{gevreg}} summarising the current model fit.}

\item{alpha}{Significance level. Default value is 0.05.}
}
\value{
An object (a list) of class \code{c("gev", "evreg")} summarising
  the new model fit (which may be the same as \code{fit}) and containing the
  following additional components
    \item{Input_fit}{The input object of the class \code{c("gev", "evreg")}.}
    \item{Note}{A message that tells if a covariate has been added or not.}
    \item{Output_fit}{A list that contains formulae for the parameter,
    and the output object of the class \code{c("gev", "evreg")} if the output fit
    is different from the input fit.}
    \item{added_covariate}{A character vector shows added covariate}
    \item{pvalue}{A data frame that contains p value with five decimal
    places of the added covariate if there is one.}
}
\description{
Add a single term to either mu, sigma, and xi based on individual p value from Wald test.
}
\details{
Add details.
}
\examples{

### Fremantle sea levels

f0 <- gevreg(SeaLevel, data = evreg::fremantle)
add1_p_mu(f0)


### Annual Maximum and Minimum Temperature

P0 <- gevreg(y = TMX1, data = PORTw[, -1])
add1_p_mu(P0)

}
