% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add1_pvalue.R
\name{add1_p_mu}
\alias{add1_p_mu}
\title{Add one possible covariate on GEV parameter using individual p value}
\usage{
add1_p_mu(fit, alpha = 0.05)
}
\arguments{
\item{fit}{A model of class "gevreg".}

\item{alpha}{Significance level. Default value is 0.05..}
}
\value{
A list which has the following components
    \item{Input_fit}{The input object of the class gevreg.}
    \item{Note}{A message that will be printed when input fit and output
    fit are the same.}
    \item{Output_fit}{A list that contains formulae for the parameter,
    and the output object of the class gevreg if output fit is different
    from the input fit.}
    \item{pvalue}{A p value with five decimal places of the covariate
    that is been added if there is one.}
}
\description{
Add a single term to either mu, sigma, and xi based on individual p value.
}
\details{
add1_p_mu(fit, alpha = 0.05)
add1_p_sigma(fit, alpha = 0.05)
add1_p_xi(fit, alpha = 0.05)

Add details.
}
\examples{

### Fremantle sea levels

f0 <- gevreg(SeaLevel, data = evreg::fremantle)
add1_p_mu(f0)

}
