% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward_AIC.R
\name{backward_AIC}
\alias{backward_AIC}
\alias{backward_AIC_mu}
\alias{backward_AIC_sigma}
\alias{backward_AIC_xi}
\title{Backward Elimination on GEV Parameter based on AIC}
\usage{
backward_AIC_mu(fit)

backward_AIC_sigma(fit)

backward_AIC_xi(fit)
}
\arguments{
\item{fit}{An object of class \code{c("gev", "evreg")} returned from
\code{\link{gevreg}} summarising the current model fit.}
}
\value{
An object (a list) of class \code{c("gev", "evreg")} summarising
  the new model fit (which may be the same as \code{fit}) and containing the
  following additional components
    \item{Input_fit}{The input object of the class \code{c("gev", "evreg")}.}
    \item{Note}{A message that tells if covariates have been dropped or not.}
    \item{Output_fit}{A list that contains formulae for the parameter,
    and the output object of the class \code{c("gev", "evreg")} if the output fit
    is different from the input fit.}
    \item{dropped_covariate}{A character vector shows dropped covariates}
    \item{AIC}{AIC values for both input model and output model if two models
    are different.}
}
\description{
Significance controlled variable selection selects variables in either
mu, sigma, and xi with backward direction based on AIC.
}
\details{
Add details.
}
\examples{

### Oxford and Worthing annual maximum temperatures

ow$year <- (ow$year - 1901) / (1980 - 1901)
ow1 <- gevreg(y = temp, data = ow[-3], mu = ~loc + year, sigma = ~loc,
xi = ~loc, sigmalink = identity)
backward_AIC_mu(ow1)


### Annual Maximum and Minimum Temperature

P6 <- gevreg(TMX1, data = PORTw[,-1], mu = ~MTMAX + AOindex + STDTMAX + STDMIN + MDTR)
P7 <- gevreg(TMX1, data = PORTw[,-1], sigma = ~MTMAX + STDTMAX + STDMIN + MDTR)
P8 <- gevreg(TMX1, data = PORTw[,-1], xi = ~MTMAX + STDTMAX + STDMIN + MDTR)
backward_AIC_mu(P6)
backward_AIC_sigma(P7)
backward_AIC_xi(P8)

}
