% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward_AIC.R
\name{backward_AIC_mu}
\alias{backward_AIC_mu}
\title{Backward Elimination on GEV Parameter based on AIC}
\usage{
backward_AIC_mu(fit)
}
\arguments{
\item{fit}{A model of class "gevreg".}
}
\value{
A list which has the following components
    \item{Input_fit}{The input object of the class gevreg.}
    \item{Note}{A message that will be printed when input fit and output
    fit are the same.}
    \item{Output_fit}{A list that contains formulae for the parameter,
    and the output object of the class gevreg if output fit is different
    from the input fit.}
    \item{AIC}{AIC values for both input model and output model if two models
    are different.}
}
\description{
Significance controlled variable selection selects variables in either
mu, sigma, and xi with backward direction based on AIC.
}
\details{
backward_AIC_mu(fit)
backward_AIC_sigma(fit)
backward_AIC_xi(fit)

Add details.
}
\examples{

### Oxford and Worthing annual maximum temperatures

ow$year <- (ow$year - 1901) / (1980 - 1901)
ow1 <- gevreg(y = temp, data = ow[-3], mu = ~loc + year, sigma = ~loc,
xi = ~loc, sigmalink = identity)
backward_AIC_mu(ow1)


### Annual Maximum and Minimum Temperature

library(extRemes)
data(PORTw)
PORTw$Year <- (PORTw$Year - min(PORTw$Year)) / (max(PORTw$Year) - min(PORTw$Year))
P6 <- gevreg(TMX1, data = PORTw, mu = ~MTMAX + AOindex + Year + STDTMAX + STDMIN + MDTR)
backward_AIC_mu(P6)

}
