% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward_gevreg.R
\name{backward_gevreg}
\alias{backward_gevreg}
\title{Backward Selection on GEV Parameter}
\usage{
backward_gevreg(fit, criterion = "pvalue", alpha = 0.05,
  do_mu = TRUE, do_sigma = FALSE, do_xi = FALSE)
}
\arguments{
\item{fit}{An object of class \code{c("gev", "evreg")} returned from
\code{\link{gevreg}} summarising the current model fit.}

\item{criterion}{Either based \code{LRT}(Likelihood ratio test),
\code{AIC}, or \code{pvalue}(by Wald test). Default criterion is \code{pvalue}.}

\item{alpha}{Significance level if criterion equals pvalue or LRT.
Default value is 0.05.}

\item{do_mu}{do backward selection on mu if \code{do_mu} equals TRUE. Default is TRUE.}

\item{do_sigma}{do backward selection on sigma if \code{do_sigma} equals TRUE. Default is FALSE.}

\item{do_xi}{do backward selection on xi if \code{do_xi} equals TRUE. Default is FALSE.}
}
\value{
An object (a list) of class \code{c("gev", "evreg")} summarising
  the new model fit (which may be the same as \code{fit}) and containing the
  following additional components
    \item{Input_fit}{The input object of the class \code{c("gev", "evreg")}.}
    \item{Note}{A message that tells if a covariate has been dropped or not.}
    \item{Output_fit}{A list that contains formulae for the parameter,
    and the output object of the class \code{c("gev", "evreg")} if the output fit
    is different from the input fit.}
    \item{dropped_covariate}{A character vector shows dropped covariate}
    \item{criterion_value}{criterion value for if both input model and output model
    are different.}
}
\description{
Significance controlled variable selection selects variables in either
mu, sigma, and xi with backward direction based on likelihood-ratio-test,
AIC, or p value from Wald test.
}
\details{
The function performs backward elimination for an object of class \code{c("gev", "evreg")}.
When \code{do_mu}, \code{do_sigma}, and \code{do_xi} all equal TRUE, the function
performs backward selection on xi first, then on sigma, and finally on mu.

Non-zero components of inital value of xi may mean that the likelihood is zero
at the starting values.  This is because for xi not equal to zero,
there is a constraint on the parameter space.  To avoid this set all
components of initial value of xi to 0, i.e. the Gumbel case.
}
\examples{
### Annual Maximum and Minimum Temperature

P6 <- gevreg(y = TMX1, data = PORTw[, -1], mu = ~MTMAX + AOindex + STDTMAX + STDMIN + MDTR)
backward_gevreg(P6)

}
