% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward_LRT.R
\name{backward_LRT_mu}
\alias{backward_LRT_mu}
\title{Backward Elimination on GEV Parameter based on Likelihood-ratio-test}
\usage{
backward_LRT_mu(fit, alpha = 0.05)
}
\arguments{
\item{fit}{A model of class "gevreg".}

\item{alpha}{Significance level. Default value is 0.05..}
}
\value{
A list which has the following components
    \item{Input_fit}{The input object of the class gevreg.}
    \item{Note}{A message that will be printed when input fit and output
    fit are the same.}
    \item{Output_fit}{A list that contains formulae for the parameter,
    and the output object of the class gevreg if output fit is different
    from the input fit.}
    \item{pvalue}{A p value based on a likelihood-ratio-test if the input fit
    and  output fit are different.}
}
\description{
Significance controlled variable selection selects variables in either
mu, sigma, and xi with backward direction based on Likelihood-ratio-test.
}
\details{
backward_LRT_mu(fit, alpha = 0.05)
backward_LRT_sigma(fit, alpha = 0.05)
backward_LRT_xi(fit, alpha = 0.05)

Add details.
}
\examples{
### Oxford and Worthing annual maximum temperatures

ow$year <- (pjn$year - 1901) / (1980 - 1901)
ow1 <- gevreg(y = temp, data = pjn[-3], mu = ~loc + year, sigma = ~loc,
xi = ~loc, sigmalink = identity)
backward_LRT_mu(ow1)

}
