% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward_pvalue.R
\name{backward_p}
\alias{backward_p}
\alias{backward_p_mu}
\alias{backward_p_sigma}
\alias{backward_p_xi}
\title{Backward Elimination on GEV Parameter using individual p value from Wald test}
\usage{
backward_p_mu(fit, alpha = 0.05)

backward_p_sigma(fit, alpha = 0.05)

backward_p_xi(fit, alpha = 0.05)
}
\arguments{
\item{fit}{An object of class \code{c("gev", "evreg")} returned from
\code{\link{gevreg}} summarising the current model fit.}

\item{alpha}{Significance level. Default value is 0.05.}
}
\value{
An object (a list) of class \code{c("gev", "evreg")} summarising
  the new model fit (which may be the same as \code{fit}) and containing the
  following additional components
    \item{Input_fit}{The input object of the class \code{c("gev", "evreg")}.}
    \item{Note}{A message that tells if a covariate has been dropped or not.}
    \item{Output_fit}{A list that contains formulae for the parameter,
    and the output object of the class \code{c("gev", "evreg")} if the output fit
    is different from the input fit.}
    \item{dropped_covariate}{A character vector shows added covariates}
    \item{pvalue}{A data frame that contains p values with five decimal
    places of the dropped covariates if there are some.}
}
\description{
Significance controlled variable selection selects variables in either
mu, sigma, and xi with backward direction based on individual p value from Wald test.
}
\details{
Add details.
}
\examples{
### Fremantle sea levels

f3 <- gevreg(y = SeaLevel, data = evreg::fremantle[-1], mu = ~Year01 + SOI)
backward_p_mu(f3)


### Annual Maximum and Minimum Temperature

P6 <- gevreg(TMX1, data = PORTw[,-1], mu = ~MTMAX + AOindex + STDTMAX + STDMIN + MDTR)
backward_p_mu(P6)

}
