% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward_pvalue.R
\name{backward_p_mu}
\alias{backward_p_mu}
\title{Backward Elimination on GEV Parameter using individual p value}
\usage{
backward_p_mu(fit, alpha = 0.05)
}
\arguments{
\item{fit}{A model of class "gevreg".}

\item{alpha}{Significance level. Default value is 0.05..}
}
\value{
A list which has the following components
    \item{Input_fit}{The input object of the class gevreg.}
    \item{Note}{A message that will be printed when input fit and output
    fit are the same.}
    \item{Output_fit}{A list that contains formulae for the parameter,
    and the output object of the class gevreg if output fit is different
    from the input fit.}
    \item{pvalue}{A data frame that contains p values with five decimal
    places of the dropped covariates in order if there are.}
}
\description{
Significance controlled variable selection selects variables in either
mu, sigma, and xi with backward direction based on individual p value.
}
\details{
backward_p_mu(fit, alpha = 0.05)
backward_p_sigma(fit, alpha = 0.05)
backward_p_xi(fit, alpha = 0.05)

Add details.
}
\examples{

### Fremantle sea levels

f3 <- gevreg(y = SeaLevel, data = evreg::fremantle[-1], mu = ~Year01 + SOI)
backward_p_mu(f3)


### Oxford and Worthing annual maximum temperatures

ow$year <- (ow$year - 1901) / (1980 - 1901)
ow1 <- gevreg(y = temp, data = ow[-3], mu = ~loc + year, sigma = ~loc,
xi = ~loc, sigmalink = identity)
backward_AIC_mu(ow1)


### Annual Maximum and Minimum Temperature

library(extRemes)
data(PORTw)
PORTw$Year <- (PORTw$Year - min(PORTw$Year)) / (max(PORTw$Year) - min(PORTw$Year))
P6 <- gevreg(TMX1, data = PORTw, mu = ~MTMAX + AOindex + Year + STDTMAX + STDMIN + MDTR)
backward_p_mu(P6)

}
