% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_pvalue.R
\name{compare_pvalue}
\alias{compare_pvalue}
\title{Likehood-ratio Test for evreg Objects}
\usage{
compare_pvalue(model1, model2)
}
\arguments{
\item{model1}{A model of class \code{c("gev", "evreg")}.}

\item{model2}{A model of class \code{c("gev", "evreg")}.}
}
\value{
A numeric number that is between 0 to 1.
}
\description{
Compare two evreg objects by using chi-squared distritbuion
}
\details{
compare_pvalue(model1, model2)


Model 1 has to be nested within model 2
}
\examples{

### Fremantle sea levels

f0 <- gevreg(SeaLevel, data = evreg::fremantle)
f1 <- gevreg(SeaLevel, data = evreg::fremantle, mu = ~ SOI)
compare_pvalue(f0, f1)

}
