% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop1_AIC.R
\name{drop1_AIC}
\alias{drop1_AIC}
\alias{drop1_AIC_mu}
\alias{drop1_AIC_sigma}
\alias{drop1_AIC_xi}
\title{Drop one possible covariate on GEV parameter based on AIC}
\usage{
drop1_AIC_mu(fit)

drop1_AIC_sigma(fit)

drop1_AIC_xi(fit)
}
\arguments{
\item{fit}{An object of class \code{c("gev", "evreg")} returned from
\code{\link{gevreg}} summarising the current model fit.}
}
\value{
An object (a list) of class \code{c("gev", "evreg")} summarising
  the new model fit (which may be the same as \code{fit}) and containing the
  following additional components
    \item{Input_fit}{The input object of the class \code{c("gev", "evreg")}.}
    \item{Note}{A message that tells if a covariate has been added or not.}
    \item{Output_fit}{A list that contains formulae for the parameter,
    and the output object of the class \code{c("gev", "evreg")} if the output fit
    is different from the input fit.}
    \item{dropped_covariate}{A character vector shows dropped covariate}
    \item{AIC}{AIC values for both input model and output model if two models
    are different.}
}
\description{
Drop a single term to either mu, sigma, and xi based on AIC.
}
\details{
Add details.
}
\examples{

### Annual Maximum and Minimum Temperature

P6 <- gevreg(TMX1, data = PORTw[,-1], mu = ~MTMAX + AOindex + STDTMAX + STDMIN + MDTR)
P7 <- gevreg(TMX1, data = PORTw[,-1], sigma = ~MTMAX + STDTMAX + STDMIN + MDTR)
P8 <- gevreg(TMX1, data = PORTw[,-1], xi = ~MTMAX + STDTMAX + STDMIN + MDTR)
drop1_AIC_mu(P6)
drop1_AIC_sigma(P7)
drop1_AIC_xi(P8)


### Oxford and Worthing annual maximum temperatures
#Parameter mu
ow$year <- (ow$year - 1901) / (1980 - 1901)
ow1 <- gevreg(y = temp, data = ow[-3], mu = ~loc + year, sigma = ~loc,
xi = ~loc, sigmalink = identity)
drop1_AIC_mu(ow1)

}
