% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop1_AIC.R
\name{drop1_AIC_mu}
\alias{drop1_AIC_mu}
\title{Drop one possible covariate on GEV parameter based on AIC}
\usage{
drop1_AIC_mu(fit)
}
\arguments{
\item{fit}{A model of class "gevreg".}
}
\value{
A list which has the following components
    \item{Input_fit}{The input object of the class gevreg.}
    \item{Note}{A message that will be printed when input fit and output
    fit are the same.}
    \item{Output_fit}{A list that contains formulae for the parameter,
    and the output object of the class gevreg if output fit is different
    from the input fit.}
    \item{AIC}{AIC values for both input model and output model if two models
    are different.}
}
\description{
Drop a single term to either mu, sigma, and xi based on AIC.
}
\details{
drop1_AIC_mu(fit)
drop1_AIC_sigma(fit)
drop1_AIC_xi(fit)

Add details.
}
\examples{

### Annual Maximum and Minimum Temperature
# Parameter mu
library(extRemes)
data(PORTw)
PORTw$Year <- (PORTw$Year - min(PORTw$Year)) / (max(PORTw$Year) - min(PORTw$Year))
P6 <- gevreg(TMX1, data = PORTw, mu = ~MTMAX + AOindex + Year + STDTMAX + STDMIN + MDTR)
drop1_AIC_mu(P6)


### Oxford and Worthing annual maximum temperatures
#Parameter mu
ow$year <- (ow$year - 1901) / (1980 - 1901)
ow1 <- gevreg(y = temp, data = ow[-3], mu = ~loc + year, sigma = ~loc,
xi = ~loc, sigmalink = identity)
drop1_AIC_mu(ow1)


}
