% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop1_LRT.R
\name{drop1_LRT}
\alias{drop1_LRT}
\alias{drop1_LRT_mu}
\alias{drop1_LRT_sigma}
\alias{drop1_LRT_xi}
\title{Drop one possible covariate on GEV parameter based on Likelihood-ratio-test}
\usage{
drop1_LRT_mu(fit, alpha = 0.05)

drop1_LRT_sigma(fit, alpha = 0.05)

drop1_LRT_xi(fit, alpha = 0.05)
}
\arguments{
\item{fit}{An object of class \code{c("gev", "evreg")} returned from
\code{\link{gevreg}} summarising the current model fit.}

\item{alpha}{Significance level. Default value is 0.05..}
}
\value{
An object (a list) of class \code{c("gev", "evreg")} summarising
  the new model fit (which may be the same as \code{fit}) and containing the
  following additional components
    \item{Input_fit}{The input object of the class \code{c("gev", "evreg")}.}
    \item{Note}{A message that tells if a covariate has been added or not.}
    \item{Output_fit}{A list that contains formulae for the parameter,
    and the output object of the class \code{c("gev", "evreg")} if the output fit
    is different from the input fit.}
    \item{dropped_covariate}{A character vector shows dropped covariate}
    \item{pvalue}{A data frame that contains p value with five decimal
    places of the Likelihood-ratio-test.}
}
\description{
Drop a single term to either mu, sigma, and xi based on likelihood-ratio-test.
}
\details{
Add details.
}
\examples{

### Annual Maximum and Minimum Temperature

P3 <- gevreg(y = TMX1, data = PORTw[, -1], mu = ~MTMAX + STDTMAX + STDMIN)
P4 <- gevreg(y = TMX1, data = PORTw[, -1], sigma = ~MTMAX + STDTMAX + STDMIN)
P5 <- gevreg(y = TMX1, data = PORTw[, -1], xi = ~MTMAX + AOindex)
drop1_LRT_mu(P3)
drop1_LRT_sigma(P4)
drop1_LRT_xi(P5)
}
