% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop1_pvalue.R
\name{drop1_p}
\alias{drop1_p}
\alias{drop1_p_mu}
\alias{drop1_p_sigma}
\alias{drop1_p_xi}
\title{Drop one possible covariate on GEV parameter based on individual p value from Wald test}
\usage{
drop1_p_mu(fit, alpha = 0.05)

drop1_p_sigma(fit, alpha = 0.05)

drop1_p_xi(fit, alpha = 0.05)
}
\arguments{
\item{fit}{An object of class \code{c("gev", "evreg")} returned from
\code{\link{gevreg}} summarising the current model fit.}

\item{alpha}{Significance level. Default value is 0.05.}
}
\value{
An object (a list) of class \code{c("gev", "evreg")} summarising
  the new model fit (which may be the same as \code{fit}) and containing the
  following additional components
    \item{Input_fit}{The input object of the class \code{c("gev", "evreg")}.}
    \item{Note}{A message that tells if a covariate has been dropped or not.}
    \item{Output_fit}{A list that contains formulae for the parameter,
    and the output object of the class \code{c("gev", "evreg")} if the output fit
    is different from the input fit.}
    \item{dropped_covariate}{A character vector shows added covariate}
    \item{pvalue}{A data frame that contains p value with five decimal
    places of the dropped covariate if there is one.}
}
\description{
Drop a single term to either mu, sigma, and xi based on individual p value from Wald test.
}
\details{
Add details.
}
\examples{

### Fremantle sea levels

f3 <- gevreg(y = SeaLevel, data = evreg::fremantle[-1], mu = ~Year01 + SOI)
drop1_p_mu(f3)


### Annual Maximum and Minimum Temperature

P3 <- gevreg(y = TMX1, data = PORTw[, -1], mu = ~MTMAX + STDTMAX + STDMIN)
P  <- gevreg(y = TMX1, data = PORTw[, -1], sigma = ~MTMAX + MTMIN + STDTMAX)
P5 <- gevreg(y = TMX1, data = PORTw[, -1], xi = ~MTMAX + AOindex)
drop1_p_mu(P3)
drop1_p_sigma(P)
drop1_p_xi(P5)
}
