% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop1_pvalue.R
\name{drop1_p_mu}
\alias{drop1_p_mu}
\title{Drop one possible covariate on GEV parameter based on individual p value}
\usage{
drop1_p_mu(fit, alpha = 0.05)
}
\arguments{
\item{fit}{A model of class "gevreg".}

\item{alpha}{Significance level. Default value is 0.05..}
}
\value{
A list which has the following components
    \item{Input_fit}{The input object of the class gevreg.}
    \item{Note}{A message that will be printed when input fit and output
    fit are the same.}
    \item{Output_fit}{A list that contains formulae for the parameter,
    and the output object of the class gevreg if output fit is different
    from the input fit.}
    \item{pvalue}{A data frame that contains p value with five decimal
    places of the dropped covariate if there is one.}
}
\description{
Drop a single term to either mu, sigma, and xi based on individual p value.
}
\details{
drop1_p_mu(fit, alpha = 0.05)
drop1_p_sigma(fit, alpha = 0.05)
drop1_p_xi(fit, alpha = 0.05)

Add details.
}
\examples{

### Fremantle sea levels

f3 <- gevreg(y = SeaLevel, data = evreg::fremantle[-1], mu = ~Year01 + SOI)
drop1_p_mu(f3)

}
