% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENNreg_holdout.R
\name{ENNreg_holdout}
\alias{ENNreg_holdout}
\title{Hyperparameter tuning for the ENNreg model using the hold-out method}
\usage{
ENNreg_holdout(
  X,
  y,
  K,
  batch = TRUE,
  val = NULL,
  nval = NULL,
  XI,
  RHO,
  nstart = 100,
  c = 1,
  lambda = 0.9,
  eps = NULL,
  nu = 1e-16,
  optimProto = TRUE,
  verbose = TRUE,
  options = list(maxiter = 1000, rel.error = 1e-04, print = 10),
  opt.rmsprop = list(batch_size = 100, epsi = 0.001, rho = 0.9, delta = 1e-08, Dtmax =
    100)
)
}
\arguments{
\item{X}{Input matrix of size n x p, where n is the number of objects and p the number of
attributes.}

\item{y}{Vector of length n containing observations of the response variable.}

\item{K}{Number of prototypes.}

\item{batch}{If TRUE (default), batch learning is used; otherwise, online learning is
used.}

\item{val}{Vector of indices of the validation instances (nval integers between 1 and n).
Needed only if \code{nval} is not provided.}

\item{nval}{Number of validation instances (needed only if \code{val} is not provided).}

\item{XI}{Vector of candidate values for hyperparameter \code{xi}.}

\item{RHO}{Vector of candidate values for hyperparameter \code{rho}.}

\item{nstart}{Number of random starts of the k-means algorithm (default: 100).}

\item{c}{Multiplicative coefficient applied to scale parameter gamma (defaut: 1).}

\item{lambda}{Parameter of the loss function (default=0.9).}

\item{eps}{Parameter of the loss function (if NULL, fixed to 0.01 times the standard
deviation of y).}

\item{nu}{Parameter of the loss function to avoid a division par zero (default=1e-16).}

\item{optimProto}{If TRUE (default), the initial prototypes are optimized.}

\item{verbose}{If TRUE (default) intermediate results are displayed.}

\item{options}{Parameters of the optimization algorithm (see \code{\link{ENNreg}}).}

\item{opt.rmsprop}{Parameters of the RMSprop algorithm (see \code{\link{ENNreg}}).}
}
\value{
A list with three components:
\describe{
\item{xi}{Optimal value of xi.}
\item{rho}{Optimal value of rho.}
\item{RMS}{Matrix of root mean squared error values}.
}
}
\description{
\code{ENNreg_holdout tunes parameters xi and rho of the ENNreg model using the
hold-out method.}
}
\details{
Either the validation set (a vector of indices), or the number nval of validation
instances must be provided. Arguments \code{options} and \code{opt.rmsprop} are passed
to function \code{\link{ENNreg}}.
}
\examples{
# Boston dataset
\donttest{
library(MASS)
X<-as.matrix(scale(Boston[,1:13]))
y<-Boston[,14]
set.seed(220322)
n<-nrow(Boston)
hold.out<-ENNreg_holdout(X,y,K=30,nval=round(n/3),XI=c(0.1,1,10),RHO=c(0.1,1,10))
hold.out$RMS
}

}
\references{
Thierry Denoeux. An evidential neural network model for regression based on random fuzzy
numbers. In "Belief functions: Theory and applications (proc. of BELIEF 2022)", pages 57-66,
Springer, 2022.

Thierry Denoeux. Quantifying prediction uncertainty in regression using random fuzzy sets: the ENNreg
model. IEEE Transactions on Fuzzy Systems, Vol. 31, Issue 10, pages 3690-3699, 2023.
}
\seealso{
\code{\link{ENNreg}}, \code{\link{ENNreg_cv}}
}
