% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evreg-package.R
\docType{package}
\name{evreg-package}
\alias{evreg}
\alias{evreg-package}
\title{evreg: Evidential Regression}
\description{
An implementation of the 'Evidential Neural Network for Regression' model recently introduced in Denoeux (2023) \doi{10.1109/TFUZZ.2023.3268200}. In this model, prediction uncertainty is quantified by Gaussian random fuzzy numbers as introduced in Denoeux (2023) \doi{10.1016/j.fss.2022.06.004}. The package contains functions for training the network, tuning hyperparameters by cross-validation or the hold-out method, and making predictions. It also contains utilities for making calculations with Gaussian random fuzzy numbers (such as, e.g., computing the degrees of belief and plausibility of an interval, or combining Gaussian random fuzzy numbers).
}
\author{
\strong{Maintainer}: Thierry Denoeux \email{tdenoeux@utc.fr} (\href{https://orcid.org/0000-0002-0660-5436}{ORCID})

}
\keyword{internal}
