% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evreg.R
\docType{package}
\name{evreg}
\alias{evreg}
\alias{evreg-package}
\title{evreg: Extreme value regression modelling}
\description{
Brief description of the package
}
\details{
More detail.
}
\seealso{
\code{\link{gevreg}} GEV generalized linear regression modelling.
}
