% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_AIC.R
\name{forward_AIC_mu}
\alias{forward_AIC_mu}
\title{AIC-based Forward Selection on GEV Parameter}
\usage{
forward_AIC_mu(fit)
}
\arguments{
\item{fit}{A model of class "gevreg".}
}
\value{
A list which has the following components
    \item{Input_fit}{The input object of the class gevreg.}
    \item{Note}{A message that will be printed when input fit and output
    fit are the same.}
    \item{Output_fit}{A list that contains formulae for the parameter,
    and the output object of the class gevreg if output fit is different
    from the input fit.}
    \item{AIC}{AIC values for both input model and output model if two models
    are different.}
}
\description{
Significance controlled variable selection selects variables in either
mu, sigma, and xi with forward direction based on AIC.
}
\details{
forward_AIC_mu(fit)
forward_AIC_sigma(fit)
forward_AIC_xi(fit)

Add details.
}
\examples{

### Fremantle sea levels

f0 <- gevreg(SeaLevel, data = evreg::fremantle)
forward_AIC_mu(f0)

}
