% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_gevreg.R
\name{forward_gevreg}
\alias{forward_gevreg}
\title{Forward Selection on GEV Parameter}
\usage{
forward_gevreg(fit, criterion = "pvalue", alpha = 0.05, do_mu = TRUE,
  do_sigma = FALSE, do_xi = FALSE)
}
\arguments{
\item{fit}{An object of class \code{c("gev", "evreg")} returned from
\code{\link{gevreg}} summarising the current model fit.}

\item{criterion}{Either based \code{LRT}(Likelihood ratio test),
\code{AIC}, or \code{pvalue}(by Wald test). Default criterion is \code{pvalue}.}

\item{alpha}{Significance level if criterion equals pvalue or LRT.
Default value is 0.05.}

\item{do_mu}{do forward selection on mu if \code{do_mu} equals TRUE. Default is TRUE.}

\item{do_sigma}{do forward selection on sigma if \code{do_sigma} equals TRUE. Default is FALSE.}

\item{do_xi}{do forward selection on xi if \code{do_xi} equals TRUE. Default is FALSE.}
}
\value{
An object (a list) of class \code{c("gev", "evreg")} summarising
  the new model fit (which may be the same as \code{fit}) and containing the
  following additional components
    \item{Input_fit}{The input object of the class \code{c("gev", "evreg")}.}
    \item{Note}{A message that tells if a covariate has been added or not.}
    \item{Output_fit}{A list that contains formulae for the parameter,
    and the output object of the class \code{c("gev", "evreg")} if the output fit
    is different from the input fit.}
    \item{added_covariate}{A character vector shows added covariate}
    \item{criterion_value}{criterion value for if both input model and output model
    are different.}
}
\description{
Significance controlled variable selection selects variables in either
mu, sigma, and xi with forward direction based on likelihood-ratio-test,
AIC, or p value from Wald test.
}
\details{
Add details.
}
\examples{
### Annual Maximum and Minimum Temperature

P0 <- gevreg(y = TMX1, data = PORTw[, -1])
forward_gevreg(P0)

}
