% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_LRT.R
\name{forward_LRT}
\alias{forward_LRT}
\alias{forward_LRT_mu}
\alias{forward_LRT_sigma}
\alias{forward_LRT_xi}
\title{Likelihood-ratio-test-based Forward Selection on GEV Parameter}
\usage{
forward_LRT_mu(fit, alpha = 0.05)

forward_LRT_sigma(fit, alpha = 0.05)

forward_LRT_xi(fit, alpha = 0.05)
}
\arguments{
\item{fit}{An object of class \code{c("gev", "evreg")} returned from
\code{\link{gevreg}} summarising the current model fit.}

\item{alpha}{Significance level. Default value is 0.05.}
}
\value{
An object (a list) of class \code{c("gev", "evreg")} summarising
  the new model fit (which may be the same as \code{fit}) and containing the
  following additional components
    \item{Input_fit}{The input object of the class \code{c("gev", "evreg")}.}
    \item{Note}{A message that tells if covariates have been added or not.}
    \item{Output_fit}{A list that contains formulae for the parameter,
    and the output object of the class \code{c("gev", "evreg")} if the output fit
    is different from the input fit.}
    \item{added_covariate}{A character vector shows added covariate}
    \item{pvalue}{A data frame that contains p values with five decimal
    places of the Likelihood-ratio-test.}
}
\description{
Significance controlled variable selection selects variables in either
mu, sigma, and xi with forward direction based on likelihood-ratio-test.
}
\details{
Add details.
}
\examples{
### Fremantle sea levels

f0 <- gevreg(SeaLevel, data = evreg::fremantle[, -1])
forward_LRT_mu(f0)


### Annual Maximum and Minimum Temperature

P0 <- gevreg(y = TMX1, data = PORTw[, -1])
forward_LRT_mu(P0)
}
