% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_LRT.R
\name{forward_LRT_mu}
\alias{forward_LRT_mu}
\title{Likelihood-ratio-test-based Forward Selection on GEV Parameter}
\usage{
forward_LRT_mu(fit, alpha = 0.05)
}
\arguments{
\item{fit}{A model of class "gevreg".}

\item{alpha}{Significance level. Default value is 0.05..}
}
\value{
A list which has the following components
    \item{Input_fit}{The input object of the class gevreg.}
    \item{Note}{A message that will be printed when input fit and output
    fit are the same.}
    \item{Output_fit}{A list that contains formulae for the parameter,
    and the output object of the class gevreg if output fit is different
    from the input fit.}
    \item{pvalue}{A p value based on a likelihood-ratio-test if the input fit
    and output fit are different.}
}
\description{
Significance controlled variable selection selects variables in either
mu, sigma, and xi with forward direction based on likelihood-ratio-test.
}
\details{
forward_LRT_mu(fit, alpha = 0.05)
forward_LRT_sigma(fit, alpha = 0.05)
forward_LRT_xi(fit, alpha = 0.05)

Add details.
}
\examples{

### Fremantle sea levels

f0 <- gevreg(SeaLevel, data = evreg::fremantle)
forward_LRT_mu(f0)

}
