% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_pvalue.R
\name{forward_p}
\alias{forward_p}
\alias{forward_p_mu}
\alias{forward_p_sigma}
\alias{forward_p_xi}
\title{Forward Selection on GEV Parameter using individual p value from Wald test}
\usage{
forward_p(fit, alpha = 0.05, do_mu = TRUE, do_sigma = FALSE,
  do_xi = FALSE)

forward_p_mu(fit, alpha = 0.05)

forward_p_sigma(fit, alpha = 0.05)

forward_p_xi(fit, alpha = 0.05)
}
\arguments{
\item{fit}{An object of class \code{c("gev", "evreg")} returned from
\code{\link{gevreg}} summarising the current model fit.}

\item{alpha}{Significance level. Default value is 0.05.}

\item{do_mu}{do forward selection on mu if \code{do_mu} equals TRUE. Default is TRUE.}

\item{do_sigma}{do forward selection on sigma after performing forward selection on mu
if \code{do_sigma} equals TRUE. Default is FALSE.}

\item{do_xi}{do forward selection on xi after doing forward selection on mu and signa
if \code{do_xi} equals TRUE. Default is FALSE.}
}
\value{
An object (a list) of class \code{c("gev", "evreg")} summarising
  the new model fit (which may be the same as \code{fit}) and containing the
  following additional components
    \item{Input_fit}{The input object of the class \code{c("gev", "evreg")}.}
    \item{Note}{A message that tells if a covariate has been added or not.}
    \item{Output_fit}{A list that contains formulae for the parameter,
    and the output object of the class \code{c("gev", "evreg")} if the output fit
    is different from the input fit.}
    \item{added_covariate}{A character vector shows added covariates}
    \item{pvalue}{A data frame that contains p values with five decimal
    places of the added covariates if there are some.}
}
\description{
Significance controlled variable selection selects variables in either
mu, sigma, and xi with forward direction based on individual p value from Wald test.
}
\details{
Add details.
}
\examples{

### Fremantle sea levels

f0 <- gevreg(SeaLevel, data = evreg::fremantle)
forward_p_mu(f0)


### Annual Maximum and Minimum Temperature

P0 <- gevreg(y = TMX1, data = PORTw[, -1])
forward_p(P0)
}
