% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evreg.R
\docType{data}
\name{fremantle}
\alias{fremantle}
\title{Annual Maximum Sea Levels at Fremantle, Western Australia}
\format{A dataframe with 86 rows and 4 columns.
  \itemize{
    \item{Column 1, \code{Year}: }{A numeric vector of years..}
    \item{Column 2, \code{Sealevel}: }{A numeric vector of annual sea level
      maxima..}
    \item{Column 3, \code{SOI}: }{A numeric vector of annual mean values of
      the Southern Oscillation Index}
    \item{Column 4, \code{Year01}: }{A shifted and scaled version of
      \code{Year} in which 1987 is mapped to 0 and 1989 to 1.
      Scaling of this kind may be useful to avoid numerical difficulties
      when using year as a covariate in a regression model.}
 }}
\source{
Coles, S. G. (2001) An Introduction to Statistical Modelling of
  Extreme Values. London: Springer.
}
\usage{
fremantle
}
\description{
Annual maximimum sea levels recorded at Fremantle, Western Australia, for
the period 1897 to 1989. Also included are annual mean values of the
Southern Oscillation Index (SOI), which is a proxy for meteorological
volatility in the Southern Hemisphere.
}
\keyword{datasets}
