% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evreg.R
\docType{data}
\name{PORTw}
\alias{PORTw}
\title{Annual Maximum and Minimum Temperature}
\format{A dataframe includes 68 observations and 9 columns
  \itemize{
    \item{\code{Year}: }{A numeric vector of the calendar year.}
    \item{\code{MTMAX}: }{A numeric vector of the means of the winter maximum temperature.}
    \item{\code{MTMIN}: }{A numeric vector of the means of the winter minimum temperature.}
    \item{\code{STDTMAX}: }{A numeric vector of the standard deviations of the winter maximum temperature.}
    \item{\code{STDMIN}: }{A numeric vector of the standard deviations of the winter minimum temperature.}
    \item{\code{TMX1}: }{A numeric vector of the annual maximum winter temperature (degrees centigrade).}
    \item{\code{TMN0}: }{A numeric vector of the annual minimum winter temperature (degrees centigrade).}
    \item{\code{MDTR}: }{A numeric vector of the mean winter diurnal temperature range.}
    \item{\code{AOindex}: }{A numeric vector of the North Atlantic Oscillation index (AOindex) from 1927 through 1995}
 }}
\usage{
PORTw
}
\description{
Annual maximum and minimum Winter temperature (degrees centigrade)
with a covariate for the North Atlantic Oscillation index from 1927 through 1995.
Data is for Winter for Port Jervis, New York (PORTw) and Spring for Sept-Iles, Quebec (SEPTsp).
}
\references{
Eric Gilleland, Richard W. Katz (2016). extRemes 2.0: An Extreme Value Analysis Package
    in R. Journal of Statistical Software, 72(8), 1-39. doi:10.18637/jss.v072.i08
}
\keyword{datasets}
