% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppreg.R
\name{ppreg}
\alias{ppreg}
\title{Point process generalized linear regression modelling}
\usage{
ppreg(y, data, p = 0.5, npy = 365, mu = ~1, sigma = ~1, xi = ~1,
  mustart, sigmastart, xistart, invmulink = identity,
  invsigmalink = identity, invxilink = identity, ...)
}
\arguments{
\item{y}{Either a numeric vector or the name of a variable in \code{data}.
\code{y} must not have any missing values.}

\item{data}{A data frame containing \code{y} and any covariates.
Neither \code{y} nor the covariates may have any missing values.}

\item{p}{Probability quantile. This is generally a number strictly between 0 and 1.}

\item{mu, sigma, xi}{Formulae (see \code{\link[stats]{formula}})
for the PP parameters \code{mu} (location), \code{sigma} (scale),
e.g. \code{mu = ~ x}. Parameter \code{xi} (shape) is holded as fixed.}
}
\description{
Regression modelling with a Point Process response distribution
and generalized linear modelling of each parameter, specified by a symbolic
description of each linear predictor and the inverse link function to be
applied to each linear predictor.
Quantile regression has been used to set a threshold for which the
probability \code{p} of threshold exceedance is approximatly constant
across different values of covarites.
}
