% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_summary.R
\name{print.evreg}
\alias{print.evreg}
\title{Print method for objects of class "evreg"}
\usage{
\method{print}{evreg}(x, digits = max(3, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{an object inheriting from class "evreg", a result of a call to
\code{\link{gevreg}}.}

\item{digits}{An integer. Used for number formatting with
\code{\link[base]{format}} and \code{\link[base:Round]{signif}}.}

\item{...}{Additional optional arguments. At present no optional
arguments are used.}
}
\value{
The argument \code{x}, invisibly, as for all
  \code{\link[base]{print}} methods.
}
\description{
\code{print} method for class "evreg".
}
\details{
Prints: the call, an indicator of convergence (\code{TRUE} only
  if \code{x$call = 0}), estimated coefficients, maximized log-likelihood,
  the number of estimated coefficients (DF) and AIC.
}
\seealso{
\code{\link{gevreg}} GEV generalized linear regression modelling.
}
