% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_summary.R
\name{summary.evreg}
\alias{summary.evreg}
\alias{print.summary.evreg}
\title{Summarizing GEV regression model fits}
\usage{
\method{summary}{evreg}(object, correlation = FALSE,
  symbolic.cor = FALSE, ...)

\method{print}{summary.evreg}(x, digits = max(5L, getOption("digits") -
  3L), symbolic.cor = x$symbolic.cor,
  signif.stars = getOption("show.signif.stars"), ...)
}
\arguments{
\item{object}{an object inheriting from class "evreg", a result of a call to
\code{\link{gevreg}}.}

\item{correlation}{logical; if TRUE, the correlation matrix of the estimated
parameters is returned and printed.}

\item{symbolic.cor}{logical. If TRUE, print the correlations in a symbolic
form (see \code{\link[stats]{symnum}}) rather than as numbers.}

\item{...}{Additional optional arguments. At present no optional
arguments are used.}

\item{digits}{An integer. Used for number formatting with
\code{\link[base]{format}} and \code{\link[base:Round]{signif}}.}

\item{signif.stars}{logical. If TRUE, 'significance stars' are printed for
each coefficient.}
}
\value{
\code{summary.glm} returns an object of class
  \code{"summary.evreg"}, a list with components
  \item{call}{the component \code{call} from \code{object}.}
  \item{coefficients}{the matrix of coefficients, standard errors, z-values
    and p-values.}
  \item{deviance}{the deviance of the fitted model, calculated as
    \code{-2 * logLik(object)}}
  \item{df.residual}{the residual degrees of freedom, calculated as
    \code{length(object$data$y) - object$df}}
  \item{aic}{the component \code{aic} from \code{object}.}

  \code{print.summary.glm} prints the call, a table of estimated
  coefficients, standard errors, the estimate/SE ratio and a two-tailed
  approximate p-value based on a standad normal reference distribution.
  Significance stars are added if \code{signif.stars = TRUE}.
  Also printed are the residual deviance and residual degrees of freedom
  and the value of AIC.
  The argument \code{x} is returned invisibly.
}
\description{
These functions are all methods for class \code{evreg} or
\code{summary.evreg} objects.  They provide similar functionality
to \code{\link[stats]{summary.glm}}.
}
\seealso{
\code{\link{gevreg}} GEV generalized linear regression modelling.
}
