% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evreg.R
\docType{data}
\name{wm}
\alias{wm}
\title{Annual Winter Maximum Wave Height}
\format{A dataframe contains 56 rows and 6 variables
  \itemize{
    \item{\code{Hs}: }{A numeric vector of the largest storm maximum value (in metres) observed
    over a given winter, which is a measure of sea surface roughness in the northern North Sea.}
    \item{\code{waterYear}: }{A numeric vector of the year in which this winter ends
    (a water year starts on 1st October)}
    \item{\code{meanNAO}: }{A numeric vector of the mean value of the North Atlantic
    Oscillation (NAO), which includes climate indice, over a winter.}
    \item{\code{maxNAO}: }{A numeric vector of the maximum value of the NAO index over a winter.}
    \item{\code{meanAO}: }{A numeric vector of the mean value of the Arctic Oscillation (AO),
    which contains climate indice, over a winter.}
    \item{\code{maxAO}: }{A numeric vector of the maximum value of the AO index over a winter.}
  }}
\usage{
wm
}
\description{
Annual Winter Maximum Wave Height in the North Sea from 1/10/1954 to 31/3/2010,
that is, the winters from 56 water years.
}
\keyword{datasets}
