% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward_LRT.R
\name{backward_LRT}
\alias{backward_LRT}
\alias{backward_LRT_mu}
\alias{backward_LRT_sigma}
\alias{backward_LRT_xi}
\title{Backward Elimination on GEV Parameter based on Likelihood-ratio-test}
\usage{
backward_LRT_mu(fit, alpha = 0.05)

backward_LRT_sigma(fit, alpha = 0.05)

backward_LRT_xi(fit, alpha = 0.05)
}
\arguments{
\item{fit}{An object of class \code{c("gev", "evreg")} returned from
\code{\link{gevreg}} summarising the current model fit.}

\item{alpha}{Significance level. Default value is 0.05.}
}
\value{
An object (a list) of class \code{c("gev", "evreg")} summarising
  the new model fit (which may be the same as \code{fit}) and containing the
  following additional components
    \item{Input_fit}{The input object of the class \code{c("gev", "evreg")}.}
    \item{Note}{A message that tells if covariates have been dropped or not.}
    \item{Output_fit}{A list that contains formulae for the parameter,
    and the output object of the class \code{c("gev", "evreg")} if the output fit
    is different from the input fit.}
    \item{dropped_covariate}{A character vector shows dropped covariates}
    \item{pvalue}{A data frame that contains p values with five decimal
    places of the Likelihood-ratio-test.}
}
\description{
Significance controlled variable selection selects variables in either
mu, sigma, and xi with backward direction based on Likelihood-ratio-test.
}
\details{
Add details.
}
\examples{
### Oxford and Worthing annual maximum temperatures

ow$year <- (pjn$year - 1901) / (1980 - 1901)
ow1 <- gevreg(y = temp, data = ow[-3], mu = ~loc + year, sigma = ~loc,
xi = ~loc, sigmalink = identity)
backward_LRT_mu(ow1)


#' ### Annual Maximum and Minimum Temperature

P3 <- gevreg(y = TMX1, data = PORTw[, -1], mu = ~MTMAX + STDTMAX + STDMIN)
P5 <- gevreg(y = TMX1, data = PORTw[, -1], xi = ~MTMAX + AOindex)
backward_LRT_mu(P3)
backward_LRT_xi(P5)

}
