% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop1_AIC.R
\name{drop1_AIC_mu}
\alias{drop1_AIC_mu}
\title{Drop one possible covariate on GEV parameter based on AIC}
\usage{
drop1_AIC_mu(fit)
}
\arguments{
\item{fit}{A model of class "gevreg".}
}
\value{
A list which has the following components
    \item{Input_fit}{The input object of the class gevreg.}
    \item{Note}{A message that will be printed when input fit and output
    fit are the same.}
    \item{Output_fit}{A list that contains formulae for the parameter,
    and the output object of the class gevreg if output fit is different
    from the input fit.}
    \item{AIC}{AIC values for both input model and output model if two models
    are different.}
}
\description{
Drop a single term to either mu, sigma, and xi based on AIC.
}
\details{
drop1_AIC_mu(fit)
drop1_AIC_sigma(fit)
drop1_AIC_xi(fit)

Add details.
}
\examples{

### Fremantle sea levels

f3 <- gevreg(y = SeaLevel, data = evreg::fremantle[-1], mu = ~Year01 + SOI)
drop1_AIC_mu(f3)

}
