% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop1_LRT.R
\name{drop1_LRT_mu}
\alias{drop1_LRT_mu}
\title{Drop one possible covariate on GEV parameter based on Likelihood-ratio-test}
\usage{
drop1_LRT_mu(fit, alpha = 0.05)
}
\arguments{
\item{fit}{A model of class "gevreg".}

\item{alpha}{Significance level. Default value is 0.05..}
}
\value{
A list which has the following components
    \item{Input_fit}{The input object of the class gevreg.}
    \item{Note}{A message that will be printed when input fit and output
    fit are the same.}
    \item{Output_fit}{A list that contains formulae for the parameter,
    and the output object of the class gevreg if output fit is different
    from the input fit.}
    \item{pvalue}{A p value based on a likelihood-ratio-test if the input fit
    and  output fit are different.}
}
\description{
Drop a single term to either mu, sigma, and xi based on likelihood-ratio-test.
}
\details{
drop1_LRT_mu(fit, alpha = 0.05)
drop1_LRT_sigma(fit, alpha = 0.05)
drop1_LRT_xi(fit, alpha = 0.05)

Add details.
}
\examples{

### Fremantle sea levels

f3 <- gevreg(y = SeaLevel, data = evreg::fremantle[-1], mu = ~Year01 + SOI)
drop1_LRT_mu(f3)

}
