% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evreg.R
\docType{package}
\name{evreg}
\alias{evreg}
\alias{evreg-package}
\title{evreg: Extreme value regression modelling}
\description{
Brief description of the package
}
\details{
The package provides functions that model non-stationary extreme values using
univariate extreme value regression modelling, while also performing variable selection
to objectively determine the best model for a given dataset. This is achieved by using
generalized linear modelling for each parameter, and by fitting models with maximum
likelihood estimates.
}
\seealso{
\code{\link{gevreg}} GEV generalized linear regression modelling.
}
